requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow
pandas==2.1.3
tabula-py==2.8.1
python-jose==3.3.0
djangorestframework-simplejwt
pyotp
jpype1
django-cors-headers
requests


